Custos

+++ backend/src/main/java/com/ico/vrp/model/Visitable.java
@@ -2,19 +2,19 @@ package main.java.com.ico.vrp.model;
 
 public class Visitable extends Location {
 
-    private final int[] timeWindow;
+    private final double[] timeWindow;
 
-    public Visitable(Location location, int[] timeWindow) {
+    public Visitable(Location location, double[] timeWindow) {
         super(location.getLatitude(), location.getLongitude());
         this.timeWindow = timeWindow;
     }
 
-    public Visitable(double latitude, double longitude, int[] timeWindow) {
+    public Visitable(double latitude, double longitude, double[] timeWindow) {
         super(latitude, longitude);
         this.timeWindow = timeWindow;
     }
 
-    public int[] getTimeWindow() {
+    public double[] getTimeWindow() {
         return timeWindow;
     }
